5.11: Blog list frontend, step11

+++ src/services/blogsService.js
@@ -11,6 +11,21 @@ const setToken = newToken => {
 
 const getAll = async (setUpdateFlag) => {
   const request = await axios.get(baseUrl, { headers: { Authorization: token } })
+
+  request.data.sort((a, b) => {
+    if (a.likes < b.likes) {
+      return 1;
+    }
+    if (a.likes > b.likes) {
+      return -1;
+    }
+
+    // names must be equal
+    return 0;
+  })
+
+
+  // console.log(request)
   return request.data
 }
 
@@ -33,8 +48,28 @@ const createBlog = async (title, author, url, setUpdateFlag) => {
 }
 
 const deleteBlog = async (id, setUpdateFlag) => {
-  const request = await axios.delete(
-    `${baseUrl}\\${id}`,
+  if (window.confirm("Do you really want to delete blog?")) {
+    const request = await axios.delete(
+      `${baseUrl}\\${id}`,
+      {
+        headers: { Authorization: token },
+        'Content-Type': 'application/json'
+      }
+    )
+    setUpdateFlag(Math.random)
+    return request.status
+  }
+  return null
+}
+
+const addLike = async (blog, setUpdateFlag) => {
+
+  const blogObject = blog
+  blogObject.likes += 1
+
+  const request = await axios.put(
+    `${baseUrl}\\${blog.id}`,
+    blogObject,
     {
       headers: { Authorization: token },
       'Content-Type': 'application/json'
@@ -45,4 +80,4 @@ const deleteBlog = async (id, setUpdateFlag) => {
   return request.data
 }
 
-export default { getAll, createBlog, deleteBlog, setToken }
+export default { getAll, createBlog, deleteBlog, addLike, setToken }
